requirements.txt
# SafarSavvy RAG - Required Dependencies Only

# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database (SQLAlchemy for SQLite - no migrations needed)
sqlalchemy==2.0.23

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pdf2image==1.16.3
PyMuPDF==1.23.4
pytesseract==0.3.10
Pillow==10.1.0

# Vector Database and Embeddings
chromadb==0.4.14
sentence-transformers>=5.1.0
numpy==1.24.3

# AI Integration (OpenRouter API)
openai>=1.12.0
httpx>=0.27.0

# Utilities
python-dotenv==1.0.0
psutil==5.9.6
python-dateutil==2.8.2
    